+++ nd4j-api/src/main/java/org/nd4j/linalg/distancefunction/EuclideanDistance.java
@@ -15,8 +15,8 @@ public class EuclideanDistance extends BaseDistanceFunction {
 	}
 
 	@Override
-	public Double apply(INDArray input) {
-		return base.distance2(input);
+	public Float apply(INDArray input) {
+		return (float) base.distance2(input);
 	}
 
 }
